[IBDA] Remove ibda.scala and move it into the LoadSliceCoreParams case class
